feat(SD-LEO-INFRA-TEST-ESTATE-HYGIENE-001): test estate hygiene — single root, trio consolidation, archive removal
VALIDATION-corrected premises (evidence 161b9738): the vitest unit project's
unanchored '**/*.test.js' include swept the orphaned legacy test/ root (CI
separately excluded it — local/CI divergence) AND tests/archived/ + docs/archived/
('**/archive/**' does not match 'archived'), whose playwright-importing .test.js
files crashed unit runs.

- FR-1: unit-project excludes += **/test/**, **/tests/archived/**, **/docs/archived/**;
  set-diff proof: collected tests/ set unchanged except the removed roots
- FR-2: legacy test/ root triaged — 31 rehomed into tests/ (7 fixed: import
  depths, stale mocks migrated to current contracts), 3 deleted (dead import,
  2 obsolete pre-migration synthesis twins); manifest at
  docs/05_testing/test-root-triage-2026-06-11.md; test/ removed
- FR-3: handoff-orchestrator trio (root .test.js 731L + handoff/ .test.js 768L
  + never-collected .spec.ts 841L) consolidated into one canonical suite
  (94 tests, 63 pass; 0 regressions — every previously-passing case still
  passes; 31 carried failures share one pre-existing preflight-mock root cause)
- FR-4: tests/archived/ (16 entries) git rm'd; vitest.config.ts.bak already
  untracked; CI tar --exclude='test/**' entries confirmed scoped to policies/
  trees (unrelated), left intact
- FR-5: skip-quarantine manifest (180 .skip/.todo across 57 files) at
  docs/05_testing/skip-quarantine-manifest-2026-06-11.md — inventory only

Descoped per VALIDATION: intelligence-layer tables EXIST live (premise refuted);
other 24 same-basename twins are mostly different modules.
