+++ tests/unit/capability-lattice.test.js
