+++ tests/unit/competitor-teardown.test.js
