+++ tests/unit/dependency-chain-validator.test.js
