+++ tests/unit/integration-discovery.test.js
