+++ tests/unit/orchestrator-child-completion.test.js
