+++ tests/unit/phase-state-machine.test.js
