+++ tests/unit/prd-metadata-consistency.test.js
