+++ tests/unit/stage-registry.test.js
