+++ tests/unit/type-aware-completion.test.js
